fix(EPv2): drop wave_config.py, inline _detect_wave_size in each consumer
The shared module was unnecessary — the HIP/TDM path already has its own
EpCfgIsValid check (worldSize <= waveSize) in C++.  Restore the original
_detect_wave_size in intranode_kernels.py and duplicate the trivial
detection inline in dispatch_combine_op.py.
