+++ python/mori/ops/dispatch_combine_v2/intranode_kernels.py
@@ -72,7 +72,25 @@ import mori.cco.device.flydsl as cco
 
 from . import flydsl_prims as P
 
-from .wave_config import WAVE
+# Wavefront size: gfx9 (MI300/MI350) = 64, gfx12 (MI400/gfx1250) = 32. Detected
+# once per process; override with MORI_WAVE_SIZE. get_warp_size(gfx1250) wrongly
+# reports 64, so key off the arch string.
+import os as _os
+
+
+def _detect_wave_size():
+    v = _os.environ.get("MORI_WAVE_SIZE")
+    if v:
+        return int(v)
+    try:
+        from mori.jit.config import detect_gpu_arch
+
+        return 32 if detect_gpu_arch().startswith("gfx12") else 64
+    except Exception:
+        return 64
+
+
+WAVE = _detect_wave_size()
 LANE_MASK = WAVE - 1
 LOG2_WAVE = WAVE.bit_length() - 1
 _BALLOT_INT = T.i64 if WAVE == 64 else T.i32
